requirements.txt
streamlit
google-generativeai
requests
tavily-python
